Refactor vehicle classes and add ShowInfo methods
Updated Car and Motorcycle constructors to remove the year parameter and added ShowInfo methods for both classes. Refactored Vehicle to remove the year property and provide a virtual ShowInfo method. Adjusted TaskOneLauncher to use the new constructors and ShowInfo methods. Deleted old Motorcycle file and replaced it with Motorcycle.cs.